+++ test/components/toolbar-flex/toolbar-flex.e2e-spec.js
@@ -19,38 +19,49 @@ describe('Flex toobar ajax tests', () => {
   });
 
   if (utils.isChrome() && utils.isCI()) {
-    it('Check ajax', async () => {
+    it('Should not visually regress', async () => {
       const flexToolbarEl = await element(by.className('no-frills'));
       await browser.driver
         .wait(protractor.ExpectedConditions.presenceOf(flexToolbarEl), config.waitsFor);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index')).toEqual(0);
-      await element(await by.id('menu-button')).click();
+      await element(await by.css('button#menu-button')).click();
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-menu-button')).toEqual(0);
       browser.driver.actions().mouseMove(element(by.css('#menu-button-popupmenu li.submenu'))).perform();
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-menu-button-submenu')).toEqual(0);
       await element(await by.css('button.btn-actions')).click();
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-more-menu')).toEqual(0);
       browser.driver.actions().mouseMove(element(by.css('button.btn-actions + div.popupmenu-wrapper li:nth-child(7)'))).perform();
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-more-menu-submenu')).toEqual(0);
 
       // shrink the page to check ajax menu button in the overflow
       browser.driver.manage().window().setSize(450, 1000);
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index')).toEqual(0);
       await element(await by.css('button.btn-actions')).click();
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-more-menu')).toEqual(0);
       browser.driver.actions().mouseMove(element(by.css('button.btn-actions + div.popupmenu-wrapper li:nth-child(2)'))).perform();
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-more-menu-overflowed-menu-button')).toEqual(0);
       browser.driver.actions().mouseMove(element(by.css('a#menu-button + div.wrapper li.submenu'))).perform();
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-more-menu-overflowed-menu-button-submenu')).toEqual(0);
-      await element(await by.id('menu-button')).click();
+      await element(await by.css('button#menu-button')).click();
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-menu-button')).toEqual(0);
       browser.driver.actions().mouseMove(element(by.css('#menu-button-popupmenu li.submenu'))).perform();
       await browser.driver.sleep(config.sleep);
+
       expect(await browser.protractorImageComparison.checkElement(flexToolbarEl, 'flextool-index-open-menu-button-submenu')).toEqual(0);
     });
   }